move helper back to strings dir
